requirements.txt
fastapi
uvicorn
openai
pinecone-client==3.0.0
pandas
tqdm
python-dotenv
